// src/enums/ui/color-palette.enum.ts
/* ColorPaletteEnum */
export enum ColorPaletteEnum {
  /* Preto */
  BLACK = "#000000",
  /* Derivado de SECONDARY_COLOR */
  DARK_SUBMENU_BACKGROUND = "#10234A", /* color-7 do Palette Generation Tool (Ant Design) */
  /* Grafite da identidade visual */
  GRAPHITE = "#161616",
  /* Azul claro da identidade visual */
  LIGHT_BLUE = "#B6E1F8",
  /* Verde da identidade visual */
  PRIMARY_COLOR = "#A2E122", /* color-6 do Palette Generation Tool (Ant Design) */
  /* Derivado de PRIMARY_COLOR */
  PRIMARY_COLOR_LIGHT = "#E9FFA1", /* color-3 do Palette Generation Tool (Ant Design) */
  /* Azul da identidade visual */
  SECONDARY_COLOR = "#1E3B6F", /* color-6 do Palette Generation Tool (Ant Design) */
  /* Derivado de SECONDARY_COLOR */
  SECONDARY_COLOR_LIGHT = "#718296", /* color-3 do Palette Generation Tool (Ant Design) */
  /* Branco */
  WHITE = "#FFFFFF"
}